Replicated structure of backupFirebase function and reverted package.json deploy:functions command
