+++ functions/scheduledFunctions/backupAuthentication.js
@@ -1,49 +1,13 @@
-const PROJECT_ID = process.env.GCLOUD_PROJECT;
-const BACKUP_BUCKET = `${PROJECT_ID}-backups`;
-const BACKUP_PATH = '/authentication/';
-//const BACKUP_SCHEDULE = 'every 1 hours synchronized';
-const BACKUP_SCHEDULE = 'every day 23:01';
-
-const admin = require('firebase-admin');
-const firebaseTools = require('firebase-tools');
-const fs = require('fs');
+const config = require('../shared/config');
+const SCHEDULE = 'every day 23:01';
 const functions = require('firebase-functions');
-const os = require('os');
-const path = require('path');
-
-// Download auth export file to the local filesystem
-const downloadAuthExport = async (filePath) => {
-  await firebaseTools.auth.export(filePath, {
-    project: PROJECT_ID,
-  });
-};
-
-// Upload local file to the backup Cloud Storage bucket
-const uploadToStorageBucket = async (localPath, fileName) => {
-  const bucket = admin.storage().bucket(BACKUP_BUCKET);
-  await bucket.upload(localPath, {
-    destination: BACKUP_PATH + fileName,
-  });
-};
-
-// Delete a file from the local filesystem
-const deleteLocalFile = (filePath) => {
-  return fs.unlinkSync(filePath);
-};
-
-// Main handler function
-const main = async () => {
-  const timestamp = (new Date()).toISOString();
-  const fileName = timestamp + '.json';
-  const tempFilePath = path.join(os.tmpdir(), fileName);
-
-  await downloadAuthExport(tempFilePath);
-  await uploadToStorageBucket(tempFilePath, fileName);
-  deleteLocalFile(tempFilePath);
-};
+const { backupAuthentication } = require('../actions/backup/authentication')(config);
 
 module.exports = functions.region('europe-west2')
-                          .pubsub
-                          .schedule(BACKUP_SCHEDULE)
-                          .timeZone('Europe/London')
-                          .onRun(main);
+  .pubsub
+  .schedule(SCHEDULE)
+  .timeZone('Europe/London')
+  .onRun(async () => {
+    const result = await backupAuthentication();
+    return result;
+  });
