+++ package.json
@@ -20,7 +20,7 @@
     "indexes:export": "firebase firestore:indexes > database/firestore.indexes.json",
     "deploy:rules": "firebase deploy --project=digital-platform-develop --only firestore:rules",
     "deploy:indexes": "firebase deploy --project=digital-platform-develop --only firestore:indexes",
-    "deploy:functions": "firebase deploy --project=digital-platform-develop --only functions:backupFirestore",
+    "deploy:functions": "firebase deploy --project=digital-platform-develop --only functions:initialiseQualifyingTest,functions:activateQualifyingTest,functions:onQualifyingTestResponseUpdate,functions:sendQualifyingTestReminders,functions:initialiseApplicationRecords,functions:initialiseMissingApplicationRecords,functions:scoreQualifyingTest,functions:onExerciseUpdate,functions:initialiseAssessments,functions:cancelAssessments,functions:testAssessmentNotification,functions:sendAssessmentRequests,functions:sendAssessmentReminders,functions:generateQualifyingTestReport,functions:generateSignInWithEmailLink,functions:exportExerciseData,functions:generateOutreachReport,functions:generateDiversityReport,functions:onPanelUpdate,functions:transferHandoverData,functions:exportApplicationContactsData,functions:generateHandoverReport,functions:generateReasonableAdjustmentsReport,functions:exportQualifyingTestResponses,functions:generateAgencyReport,functions:logEvent,functions:onApplicationUpdate,functions:onCandidatePersonalDetailsUpdate,functions:exportApplicationEligibilityIssues,functions:flagApplicationIssuesForExercise,functions:exportApplicationCharacterIssues, functions:backupFirestore, functions:backupAuthentication",
     "deploy:storage-rules": "firebase deploy --project=digital-platform-develop --only storage:rules",
     "deploy:realtime-rules": "firebase deploy --project=digital-platform-develop --only database",
     "staging:deploy:rules": "firebase deploy --project=digital-platform-staging --only firestore:rules",
